implement the rest of endpoints and the main business logic

+++ app/models/material.py
@@ -1,28 +1,15 @@
-from sqlalchemy import Column, Integer, String, Enum, Float
+from sqlalchemy import Column, String, Enum, Float, BigInteger
 from sqlalchemy.ext.declarative import declarative_base
-
 from app.schemas.material import MaterialCategory
 
 Base = declarative_base()
 
-# class MaterialCategory(enum.Enum): # todo smazat toto, pouzivam ten ze schemas/material.py
-#     FABRIC = "FABRIC"
-#     LEATHER = "LEATHER"
-#     WOOD = "WOOD"
-#     METAL = "METAL"
-#     PLASTIC = "PLASTIC"
-#     PAPER = "PAPER"
-#     COATING = "COATING"
-#     UNCATEGORIZED = "UNCATEGORIZED"
-
 class Material(Base):
     __tablename__ = "materials"
 
-    id = Column(Integer, primary_key=True, index=True) # todo pridat jeste nejake indexy?
+    id = Column(BigInteger, primary_key=True, index=True) # todo pridat jeste nejake indexy?
     name = Column(String)
     category = Column(Enum(MaterialCategory))
-
-     # todo image paths neukladam, budou vzdy stejne a jen se vezme server ID a za to specular a nonspecular treba
 
     characteristics_brightness = Column(Float)
     characteristics_color_vibrancy = Column(Float)
